backend/requirements.txt
# Flask 相关
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.10.1
Flask-Mail==0.10.0

# 序列化/反序列化
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
flask-marshmallow==1.3.0

# 数据库
PyMySQL==1.1.1
redis==5.2.1

# 文档处理
python-docx==1.1.2
openpyxl==3.1.5
python-pptx
PyMuPDF==1.25.1
pdf2docx==0.5.8
pypdf==5.2.0
pdfdeal==1.0.2
pdfkit==1.0.0
docx2pdf==0.1.8
pandas==2.2.3

# 图像处理
opencv-python-headless>=4.10.0.84
pytesseract

# 几何处理
Shapely==2.0.7

# AI & 翻译相关
tiktoken
BabelDOC==0.5.22
# 其他工具
python-dotenv==1.0.1
pymdown-extensions
openai==1.65.3